fix issues iTetris Results
Former-commit-id: 9afd073eb5c57cc7c138a8efb541febce525865e

+++ ns-3.20/src/iTETRIS-Results/model/iTETRIS-Results.h
@@ -49,12 +49,12 @@ namespace ns3
 
 
     struct PDRdata{
-        uint32_t countTx[N_STEPS_METRIC] = { 0 };
-        uint32_t countRx[N_STEPS_METRIC] = { 0 };
+        uint32_t countTx[N_STEPS_METRIC] = {0};
+        uint32_t countRx[N_STEPS_METRIC] = {0};
     };
     struct NARdata{
-        std::map <int,double> detectedVehicles;
-        std::map <int,double> totalVehicles;
+        std::map <int,int> detectedVehicles;
+        std::map <int,int> totalVehicles;
     };
     struct NIRdata{
         double countRx[N_STEPS_METRIC] = { 0 };
